fix float16 problem

+++ inference.py
@@ -68,7 +68,7 @@ def param2stroke(param, H, W, meta_brushes):
     # Dilation and erosion are used for foregrounds and alphas respectively to prevent artifacts on stroke borders.
     # print(torch.cuda.memory_summary())
     alphas = morphology.erosion(
-        (brush > 0).to(torch.float16).repeat(1, 3, 1, 1))
+        (brush > 0).to(torch.float32).repeat(1, 3, 1, 1))
     # print(torch.cuda.memory_summary())
     # Give color to foreground strokes.
     # color_map = torch.cat(
@@ -121,10 +121,10 @@ def param2img_parallel(param, decision, meta_brushes, cur_canvas):
                                     patch_size_y // 4, patch_size_y // 4, 0, 0, 0, 0])
     foregrounds = torch.zeros(
         param.shape[0], 3, patch_size_y, patch_size_x,
-        dtype=torch.float16, device=cur_canvas.device)
+        dtype=torch.float32, device=cur_canvas.device)
     alphas = torch.zeros(
         param.shape[0], 3, patch_size_y, patch_size_x,
-        dtype=torch.float16, device=cur_canvas.device)
+        dtype=torch.float32, device=cur_canvas.device)
     # print(torch.cuda.memory_summary())
     valid_foregrounds, valid_alphas = param2stroke(param[decision, :], patch_size_y, patch_size_x, meta_brushes)
     foregrounds[decision, :, :, :] = valid_foregrounds
@@ -210,7 +210,7 @@ def read_img(img_path, img_type='RGB', h=None, w=None):
     img = np.array(img)
     if img.ndim == 2:
         img = np.expand_dims(img, axis=-1)
-    img = img.transpose((2, 0, 1)).astype(np.float16) / 255
+    img = img.transpose((2, 0, 1)).astype(np.float32) / 255
     img = torch.from_numpy(img).unsqueeze(0)
     return img
 
@@ -295,7 +295,7 @@ def main(input_path, model_path, output_dir, resize_h=None, resize_w=None):
                 shape_param[index: last, ...] = output_param
                 stroke_decision[index: last, ...] = output_decision
             # shape_param, stroke_decision = net_g(img_patch, result_patch)
-            shape_param = shape_param.contiguous().to(torch.float16)
+            shape_param = shape_param.contiguous().to(torch.float32)
             stroke_decision = stroke_decision.contiguous()
             stroke_decision = network.SignWithSigmoidGrad.apply(stroke_decision)
 
@@ -351,7 +351,7 @@ def main(input_path, model_path, output_dir, resize_h=None, resize_w=None):
             shape_param[index: last, ...] = output_param
             stroke_decision[index: last, ...] = output_decision
         # shape_param, stroke_decision = net_g(img_patch, result_patch)
-        shape_param = shape_param.contiguous().to(torch.float16)
+        shape_param = shape_param.contiguous().to(torch.float32)
         stroke_decision = stroke_decision.contiguous()
 
         grid = shape_param[:, :, :2].view(img_patch.shape[0] * stroke_num, 1, 1, 2).contiguous()
